requirements.txt
streamlit==1.46.1
pandas==2.3.0
openpyxl==3.1.5
plotly==6.2.0
streamlit-aggrid==1.1.6
sqlalchemy
psycopg2-binary==2.9.10
bcrypt==4.3.0
python-dotenv==1.0.1
streamlit-authenticator==0.3.3
streamlit-oauth==0.1.14
oauth2client==4.1.3
gspread==6.2.1
dotenv==0.9.9
fpdf2==2.8.4
matplotlib==3.10.3
toml==0.10.2
sendgrid==6.12.5
google-auth-oauthlib==1.2.2
altair==5.5.0